pyproject.toml
[project]
name = "hearthstonecardgenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "networkx>=3.6.1",
    "python-dotenv>=1.2.2",
    "pyvis>=0.3.2",
]
